fix: split Political Sight Widgetbook stories under NCL cap (#4406)
catalog_panel_map_stories.dart exceeded the 1000 non-comment-line repo lint, which failed Quality. Move the Sight use cases into a domain-named sibling part.
